feature(discussions): converts discussion reply JS into an AMD module
Discussion reply javascript is now an async module loaded from
object/discussion_reply view, which will make discussion reply
editing work outside of the discussion view page.
